requirements.txt
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
certifi==2023.7.22
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7
decorator==5.1.1
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
djangorestframework-gis==1.0
exceptiongroup==1.1.3
executing==1.2.0
Fiona==1.9.4.post1
GeoAlchemy2==0.14.1
geopandas==0.14.0
greenlet==2.0.2
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.2
Markdown==3.4.4
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.26.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
pip==23.2.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.0
psycopg2==2.9.7
pure-eval==0.2.2
Pygments==2.16.1
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==305.1
pyzmq==25.1.0
setuptools==68.0.0
shapely==2.0.1
six==1.16.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.2
traitlets==5.10.1
typing_extensions==4.8.0
tzdata==2023.3
unicodecsv==0.14.1
wcwidth==0.2.6
wheel==0.41.2
zipp==3.17.0
